+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Current
 
-- Add [Aave v3 deployer support](https://github.com/aave/aave-v3-deploy) 
-- Add Solidity library linking support for Hardhat-based deployments 
+- Add [Aave v3 deployer support](https://github.com/aave/aave-v3-deploy) in`eth_defi.aave_v3.deployer` module
+- Add Solidity library linking support for Hardhat-based deployments in `eth_defi.abi.get_linked_contract`
 
 # 0.18.4